treenet_api/requirements.txt
certifi==2018.1.18
chardet==3.0.4
Django==1.11.10
django-rest-framework-mongoengine==3.3.1
djangorestframework==3.7.7
funcsigs==1.0.2
idna==2.6
mock==2.0.0
mongoengine==0.15.0
pbr==3.1.1
pyaml==17.12.1
pymongo==2.7.2
pytz==2018.3
PyYAML==3.12
requests==2.18.4
six==1.11.0
urllib3==1.22
